--- app/core/navs/documentation.js
@@ -1,570 +1,0 @@
-export const documentationNav = [
-  {
-    title: {
-      title: "Introduction",
-      iconPath: "/img/introduction.svg",
-      iconDarkPath: "/img/introduction-white.svg",
-    },
-    pages: [
-      {
-        title: "Get Started with Blitz",
-        sidebar_label: "Get Started",
-        href: "/docs/get-started",
-      },
-      {
-        title: "Tutorial",
-        sidebar_label: "Tutorial",
-        href: "/docs/tutorial",
-      },
-      {
-        title: "What is Next.js?",
-        sidebar_label: "What is Next.js?",
-        href: "/docs/what-is-nextjs",
-      },
-      {
-        title: "Why use Blitz instead of Next.js?",
-        sidebar_label: "Why Blitz instead of Next.js?",
-        href: "/docs/why-blitz",
-      },
-      {
-        title: "Tradeoffs",
-        sidebar_label: "Tradeoffs",
-        href: "/docs/tradeoffs",
-      },
-      {
-        title: "Want Free Stickers?",
-        sidebar_label: "Free Stickers",
-        displayUrl: "/stickers",
-        titleAlign: "center",
-        hideFooter: true,
-        href: "/docs/stickers",
-      },
-    ],
-  },
-  {
-    title: {
-      title: "Community",
-      iconPath: "/img/people-purple.svg",
-      iconDarkPath: "/img/people-white.svg",
-    },
-    pages: [
-      {
-        title: "How the Community Operates",
-        sidebar_label: "How the Community Operates",
-        href: "/docs/how-the-community-operates",
-      },
-      {
-        id: "manifesto",
-        title: "The Blitz.js Manifesto",
-        sidebar_label: "Manifesto",
-        href: "/docs/manifesto",
-      },
-      {
-        title: "Community History",
-        sidebar_label: "History",
-        href: "/docs/community-history",
-      },
-      {
-        title: "How to Contribute",
-        sidebar_label: "How to Contribute",
-        href: "/docs/contributing",
-      },
-      {
-        title: "❤️ Blitz Maintainers",
-        sidebar_label: "Being a Maintainer",
-        href: "/docs/maintainers",
-      },
-      {
-        id: "code-of-conduct",
-        title: "The Blitz Community Code of Conduct",
-        sidebar_label: "Code of Conduct",
-        href: "/docs/code-of-conduct",
-      },
-      {
-        title: "Help Us Translate Blitz Docs",
-        sidebar_label: "Doc Translations",
-        href: "/docs/translations",
-      },
-    ],
-  },
-  {
-    title: {
-      title: "Basics",
-      iconPath: "/img/basics.svg",
-      iconDarkPath: "/img/basics-white.svg",
-    },
-    pages: [
-      {
-        id: "file-structure",
-        title: "File Structure",
-        sidebar_label: "File Structure",
-        href: "/docs/file-structure",
-      },
-      {
-        title: "Root <App> Component",
-        sidebar_label: "<App>",
-        href: "/docs/app-component",
-      },
-      {
-        title: "Built-In CSS Support",
-        sidebar_label: "Built-In CSS Support",
-        href: "/docs/css",
-      },
-      {
-        title: "Image Component and Image Optimization",
-        sidebar_label: "Image Optimization",
-        href: "/docs/image-optimization",
-      },
-      {
-        title: "Static Files & Images",
-        sidebar_label: "Static Files & Images",
-        href: "/docs/static-files",
-      },
-      {
-        title: "Environment Variables",
-        sidebar_label: "Environment Variables",
-        href: "/docs/environment-variables",
-      },
-      {
-        title: "Error Handing",
-        sidebar_label: "Error Handling",
-        href: "/docs/error-handling",
-      },
-      {
-        title: "Testing",
-        sidebar_label: "Testing",
-        href: "/docs/testing",
-      },
-    ],
-  },
-  {
-    title: {
-      title: "Pages",
-      iconPath: "/img/pages.svg",
-      iconDarkPath: "/img/pages-white.svg",
-    },
-    pages: [
-      {
-        title: "Pages",
-        sidebar_label: "Pages",
-        href: "/docs/pages",
-      },
-      {
-        title: "Redirects",
-        sidebar_label: "Redirects",
-        href: "/docs/redirects",
-      },
-      {
-        title: "Error Pages",
-        sidebar_label: "Error Pages",
-        href: "/docs/error-pages",
-      },
-      {
-        title: "<Head> Component",
-        sidebar_label: "<Head>",
-        href: "/docs/head-component",
-      },
-      {
-        title: "<Document>",
-        sidebar_label: "<Document>",
-        href: "/docs/document-component",
-      },
-      {
-        title: "Preview Mode",
-        sidebar_label: "Preview Mode",
-        href: "/docs/preview-mode",
-      },
-      {
-        title: "Static HTML Export",
-        sidebar_label: "Static HTML Export",
-        href: "/docs/static-html-export",
-      },
-      {
-        title: "Code Splitting",
-        sidebar_label: "Code Splitting",
-        href: "/docs/code-splitting",
-      },
-      {
-        title: "getStaticProps API",
-        sidebar_label: "getStaticProps API",
-        href: "/docs/get-static-props",
-      },
-      {
-        title: "getStaticPaths API",
-        sidebar_label: "getStaticPaths API",
-        href: "/docs/get-static-paths",
-      },
-      {
-        title: "getServerSideProps API",
-        sidebar_label: "getServerSideProps API",
-        href: "/docs/get-server-side-props",
-      },
-    ],
-  },
-  {
-    title: {
-      title: "Routing",
-      iconPath: "/img/routing.svg",
-      iconDarkPath: "/img/routing-white.svg",
-    },
-    pages: [
-      {
-        title: "File-Based Routing",
-        sidebar_label: "File-Based Routing",
-        href: "/docs/routing",
-      },
-      {
-        title: "Routing Conventions",
-        sidebar_label: "Conventions",
-        href: "/docs/routing-conventions",
-      },
-      {
-        title: "Internationalized Routing",
-        sidebar_label: "Internationalized Routing",
-        href: "/docs/i18n-routing",
-      },
-      {
-        title: "<Link>",
-        sidebar_label: "<Link>",
-        href: "/docs/link",
-      },
-      {
-        title: "URL Params & Query",
-        sidebar_label: "URL Params & Query",
-        href: "/docs/route-params-query",
-      },
-      {
-        title: "Router",
-        sidebar_label: "Router",
-        href: "/docs/router",
-      },
-      {
-        title: "API Routes",
-        sidebar_label: "API Routes",
-        href: "/docs/api-routes",
-      },
-    ],
-  },
-  {
-    title: {
-      title: "Auth",
-      iconPath: "/img/shield-purple.svg",
-      iconDarkPath: "/img/shield-white.svg",
-    },
-    pages: [
-      {
-        title: "Auth Overview",
-        sidebar_label: "Overview",
-        href: "/docs/auth",
-      },
-      {
-        title: "Session Management",
-        sidebar_label: "Session Management",
-        href: "/docs/session-management",
-      },
-      {
-        title: "Authorization & Security",
-        sidebar_label: "Authorization & Security",
-        href: "/docs/authorization",
-      },
-      {
-        title: "Auth Hooks & Utilities",
-        sidebar_label: "Hooks & Utilities",
-        href: "/docs/auth-utils",
-      },
-      {
-        title: "Third Party Login with Passport.js",
-        sidebar_label: "Third Party Login w/Passport.js",
-        href: "/docs/passportjs",
-      },
-      {
-        title: "How To Impersonate Other Users",
-        sidebar_label: "How To Impersonate Other Users",
-        href: "/docs/impersonation",
-      },
-    ],
-  },
-  {
-    title: {
-      title: "Database",
-      iconPath: "/img/database.svg",
-      iconDarkPath: "/img/database-white.svg",
-    },
-    pages: [
-      {
-        title: "Database Overview",
-        sidebar_label: "Overview",
-        href: "/docs/database-overview",
-      },
-      {
-        title: "Run Postgres Locally",
-        sidebar_label: "Run Postgres Locally",
-        href: "/docs/postgres",
-      },
-      {
-        title: "Database Seeds",
-        sidebar_label: "Seeds",
-        href: "/docs/database-seeds",
-      },
-      {
-        title: "Prisma Utilities",
-        sidebar_label: "Prisma Utilities",
-        href: "/docs/prisma",
-      },
-      {
-        title: "Using Fauna with Blitz.js",
-        sidebar_label: "Fauna",
-        href: "/docs/fauna",
-      },
-    ],
-  },
-  {
-    title: {
-      title: "Queries & Mutations",
-      iconPath: "/img/queries.svg",
-      iconDarkPath: "/img/queries-white.svg",
-    },
-    pages: [
-      {
-        title: "Query Resolvers",
-        sidebar_label: "Query Resolvers",
-        href: "/docs/query-resolvers",
-      },
-      {
-        title: "Using Queries",
-        sidebar_label: "Use Queries",
-        href: "/docs/query-usage",
-      },
-      {
-        title: "Mutation Resolvers",
-        sidebar_label: "Mutation Resolvers",
-        href: "/docs/mutation-resolvers",
-      },
-      {
-        title: "Using Mutations",
-        sidebar_label: "Use Mutations",
-        href: "/docs/mutation-usage",
-      },
-      {
-        title: "Client Utilities",
-        sidebar_label: "Client Utilities",
-        href: "/docs/resolver-client-utilities",
-      },
-      {
-        title: "Server Utilities",
-        sidebar_label: "Server Utilities",
-        href: "/docs/resolver-server-utilities",
-      },
-      {
-        title: "useQuery",
-        sidebar_label: "useQuery",
-        href: "/docs/use-query",
-      },
-      {
-        title: "usePaginatedQuery",
-        sidebar_label: "usePaginatedQuery",
-        href: "/docs/use-paginated-query",
-      },
-      {
-        title: "useInfiniteQuery",
-        sidebar_label: "useInfiniteQuery",
-        href: "/docs/use-infinite-query",
-      },
-      {
-        title: "useMutation",
-        sidebar_label: "useMutation",
-        href: "/docs/use-mutation",
-      },
-    ],
-  },
-  {
-    title: {
-      title: "Backend Architecture",
-      iconPath: "/img/mutations.svg",
-      iconDarkPath: "/img/mutations-white.svg",
-    },
-    pages: [
-      {
-        title: "Background Processing with Quirrel",
-        sidebar_label: "Background Processing with Quirrel",
-        href: "/docs/background-processing-with-quirrel",
-      },
-      {
-        title: "HTTP Middleware",
-        sidebar_label: "HTTP Middleware",
-        href: "/docs/middleware",
-      },
-      {
-        title: "Custom Server",
-        sidebar_label: "Custom Server",
-        href: "/docs/custom-server",
-      },
-    ],
-  },
-  {
-    title: {
-      title: "Deploying to Production",
-      iconPath: "/img/deploying-to-production.svg",
-      iconDarkPath: "/img/deploying-to-production-white.svg",
-    },
-    pages: [
-      {
-        title: "Deploy to a Server on Render.com",
-        sidebar_label: "To Render.com",
-        href: "/docs/deploy-render",
-      },
-      {
-        title: "Deploy to a Server on Heroku",
-        sidebar_label: "To Heroku",
-        href: "/docs/deploy-heroku",
-      },
-      {
-        title: "Deploy Serverless to Vercel",
-        sidebar_label: "To Vercel",
-        href: "/docs/deploy-vercel",
-      },
-    ],
-  },
-  {
-    title: {
-      title: "Recipes",
-      iconPath: "/img/recipe-purple.svg",
-      iconDarkPath: "/img/recipe-white.svg",
-    },
-    pages: [
-      {
-        title: "Using Recipes",
-        sidebar_label: "Using Recipes",
-        href: "/docs/using-recipes",
-      },
-      {
-        title: "Writing Your Own Recipes",
-        sidebar_label: "Writing Recipes",
-        href: "/docs/writing-recipes",
-      },
-    ],
-  },
-  {
-    title: {
-      title: "Configuration",
-      iconPath: "/img/config-purple.svg",
-      iconDarkPath: "/img/config-white.svg",
-    },
-    pages: [
-      {
-        title: "blitz.config.js",
-        sidebar_label: "blitz.config.js",
-        href: "/docs/blitz-config",
-      },
-      {
-        title: "Custom Webpack Config",
-        sidebar_label: "Webpack Config",
-        href: "/docs/webpack-config",
-      },
-      {
-        title: "Custom PostCSS Config",
-        sidebar_label: "PostCSS Config",
-        href: "/docs/postcss-config",
-      },
-      {
-        id: "rpc-specification",
-        title: "RPC Specification",
-        sidebar_label: "RPC Specification",
-        href: "/docs/rpc-specification",
-      },
-      {
-        title: "Measuring Performance",
-        sidebar_label: "Measuring Performance",
-        href: "/docs/measuring-performance",
-      },
-    ],
-  },
-  {
-    title: {
-      title: "CLI",
-      iconPath: "/img/terminal-purple.svg",
-      iconDarkPath: "/img/terminal-white.svg",
-    },
-    pages: [
-      {
-        title: "Blitz CLI Overview",
-        sidebar_label: "Overview",
-        href: "/docs/cli-overview",
-      },
-      {
-        title: "blitz new",
-        sidebar_label: "blitz new",
-        href: "/docs/cli-new",
-      },
-      {
-        title: "blitz dev",
-        sidebar_label: "blitz dev",
-        href: "/docs/cli-dev",
-      },
-      {
-        title: "blitz start",
-        sidebar_label: "blitz start",
-        href: "/docs/cli-start",
-      },
-      {
-        title: "blitz build",
-        sidebar_label: "blitz build",
-        href: "/docs/cli-build",
-      },
-      {
-        title: "blitz export",
-        sidebar_label: "blitz export",
-        href: "/docs/cli-export",
-      },
-      {
-        title: "blitz db",
-        sidebar_label: "blitz db",
-        href: "/docs/cli-db",
-      },
-      {
-        title: "blitz prisma",
-        sidebar_label: "blitz prisma",
-        href: "/docs/cli-prisma",
-      },
-      {
-        title: "blitz generate",
-        sidebar_label: "blitz generate",
-        href: "/docs/cli-generate",
-      },
-      {
-        title: "blitz console",
-        sidebar_label: "blitz console",
-        href: "/docs/cli-console",
-      },
-      {
-        title: "blitz install",
-        sidebar_label: "blitz install",
-        href: "/docs/cli-install",
-      },
-      {
-        title: "blitz autocomplete",
-        sidebar_label: "blitz autocomplete",
-        href: "/docs/cli-autocomplete",
-      },
-      {
-        title: "blitz routes",
-        sidebar_label: "blitz routes",
-        href: "/docs/cli-routes",
-      },
-    ],
-  },
-  {
-    title: {
-      title: "Templates",
-      iconPath: "/img/template-purple.svg",
-      iconDarkPath: "/img/template-white.svg",
-    },
-    pages: [
-      {
-        title: "Blitz Templating Language",
-        sidebar_label: "Blitz Templates",
-        href: "/docs/templates",
-      },
-    ],
-  },
-]
