+++ package.json
@@ -2,14 +2,13 @@
   "name": "blitz-site-v2",
   "version": "1.0.0",
   "scripts": {
-    "dev": "yarn gen-nav && blitz dev",
+    "dev": "blitz dev",
     "start": "blitz start",
-    "build": "yarn gen-nav && blitz build",
+    "build": "blitz build",
     "lint": "yarn lint:code && yarn lint:docs",
     "lint:code": "eslint --ignore-path .gitignore --ext .js,.ts,.tsx .",
     "lint:docs": "alex app/pages/docs/**/*.mdx",
     "english-slugify": "node scripts/english-slugify.js",
-    "gen-nav": "node scripts/generate-documentation-nav.js && prettier -w app/core/navs/*",
     "postinstall": "husky install"
   },
   "lint-staged": {
